Extract default import replacement functions from named-exports-only for use in exports-at-eof
